Added a ResetButton component and gave ImgButton easy props overriding from components that use it as a base
