+++ src/components/StudentView.js
@@ -1,14 +1,14 @@
 import React, { Component } from "react";
 import "../css/simulation.css";
 import SimViewer from "./SimViewer";
-import UIBar from "./UIBar/UIBar";
+import UIBar from "./UI/UIBar";
 import back from "../assets/back.png";
 import forward from "../assets/forward.png";
 import reset from "../assets/reset.png";
 import help from "../assets/help.png";
-import ElementButton from "./ElementButton";
 import styled from "styled-components";
-import ImgButton from "./ImgButton";
+import ImgButton from "./UI/ImgButton";
+import ResetButton from "./UI/ResetButton";
 
 const Container = styled.div`
   display: flex;
@@ -47,10 +47,10 @@ class StudentView extends Component {
   render() {
     const Controls = (
       <React.Fragment>
-        <ImgButton key="back" icon id="back" src={back} />
-        <ImgButton key="forward" icon id="forward" src={forward} />
-        <ImgButton key="reset" icon id="reset" src={reset} />
-        <ImgButton key="help" icon id="help" src={help} />
+        <ImgButton key="back" id="back" src={back} />
+        <ImgButton key="forward" id="forward" src={forward} />
+        <ResetButton key="reset" id="reset" src={reset} />
+        <ImgButton key="help" id="help" src={help} />
       </React.Fragment>
     );
     return (
